fix: Translation system & organized file structure
