+++ metadata.json
@@ -6,6 +6,7 @@
 	"gettext-domain": "Noiseclapper",
 	"settings-schema": "org.gnome.shell.extensions.noiseclapper",
 	"shell-version": [
+		"45",
 		"46"
 	]
 }
